Moved to stockmedia package
